requirements.txt
Pillow==5.2.0
click==7.0